simplified DrivetrainMovementCommand for newbies tomorrow
